+++ src/main/java/de/mossgrabers/convertwithmoss/core/ConverterBackend.java
@@ -43,6 +43,7 @@ import de.mossgrabers.convertwithmoss.format.bitwig.BitwigMultisampleCreator;
 import de.mossgrabers.convertwithmoss.format.bitwig.BitwigMultisampleDetector;
 import de.mossgrabers.convertwithmoss.format.bliss.BlissCreator;
 import de.mossgrabers.convertwithmoss.format.bliss.BlissDetector;
+import de.mossgrabers.convertwithmoss.format.cmi3.FairlightCmi3Creator;
 import de.mossgrabers.convertwithmoss.format.cmi3.FairlightCmi3Detector;
 import de.mossgrabers.convertwithmoss.format.decentsampler.DecentSamplerCreator;
 import de.mossgrabers.convertwithmoss.format.decentsampler.DecentSamplerDetector;
@@ -250,6 +251,7 @@ public class ConverterBackend
         this.creators.add (new EmulatorXCreator (notifier));
         this.creators.add (new TonverkMultiCreator (notifier));
         this.creators.add (new TonverkPresetCreator (notifier));
+        this.creators.add (new FairlightCmi3Creator (notifier));
         this.creators.add (new DirectWaveCreator (notifier));
         this.creators.add (new KMPCreator (notifier));
         this.creators.add (new KorgmultisampleCreator (notifier));